requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
bcrypt==4.3.0
binaryornot==0.4.4
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==1.7.3
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.9
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-time==0.2.0
libpass==1.9.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
orjson==3.10.16
platformdirs==4.3.7
poyo==0.5.0
pre-commit==2.21.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9
shellingham==1.5.4
simple-term-menu==1.6.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
termcolor==1.1.0
text-unidecode==1.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.0
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
